requirements.txt
pytest
pre-commit
colorlog
requests
tqdm
click
sentinelhub
python-dotenv
Pillow
google-cloud-storage
matplotlib
rasterio
ipykernel
